+++ dimas/src/agent.rs
@@ -16,6 +16,7 @@ use crate::com::subscriber::{Subscriber, SubscriberBuilder};
 use crate::timer::{Timer, TimerBuilder};
 use std::{
 	marker::PhantomData,
+	ops::Deref,
 	sync::{Arc, RwLock},
 	time::Duration,
 };
@@ -24,20 +25,18 @@ use tokio::time::sleep;
 use zenoh::liveliness::LivelinessToken;
 // endregion:	--- modules
 
-// region:		--- types
-//type AgentProps<P> = std::fmt::Debug + Send + Sync + Unpin + 'static;
-// endregion:	--- types
-
 // region:		--- Agent
-/// implementation of an agent
+/// Agent
 pub struct Agent<'a, P>
 where
 	P: std::fmt::Debug + Send + Sync + Unpin + 'static,
 {
+	com: Arc<Communicator>,
 	pd: PhantomData<&'a P>,
+	// The agents property structure
+	props: Arc<RwLock<P>>,
 	#[cfg(feature = "liveliness")]
 	liveliness: bool,
-	com: Arc<Communicator>,
 	// an optional liveliness token
 	#[cfg(feature = "liveliness")]
 	liveliness_token: RwLock<Option<LivelinessToken<'a>>>,
@@ -59,8 +58,17 @@ where
 	// registered timer
 	#[cfg(feature = "timer")]
 	timers: Arc<RwLock<Vec<Timer<P>>>>,
-	// The agents propertie structure
-	props: Arc<RwLock<P>>,
+}
+
+impl<'a, P> Deref for Agent<'a, P>
+where
+	P: std::fmt::Debug + Send + Sync + Unpin + 'static,
+{
+	type Target = Arc<RwLock<P>>;
+
+	fn deref(&self) -> &Self::Target {
+		&self.props
+	}
 }
 
 impl<'a, P> Agent<'a, P>
@@ -73,9 +81,10 @@ where
 		let pd = PhantomData {};
 		Self {
 			pd,
+			com,
+			props: Arc::new(RwLock::new(properties)),
 			#[cfg(feature = "liveliness")]
 			liveliness: false,
-			com,
 			#[cfg(feature = "liveliness")]
 			liveliness_token: RwLock::new(None),
 			#[cfg(feature = "liveliness")]
@@ -90,35 +99,47 @@ where
 			queries: Arc::new(RwLock::new(Vec::new())),
 			#[cfg(feature = "timer")]
 			timers: Arc::new(RwLock::new(Vec::new())),
-			props: Arc::new(RwLock::new(properties)),
 		}
 	}
 
 	/// get the agents uuid
+	#[must_use]
 	pub fn uuid(&self) -> String {
 		self.com.uuid()
 	}
 
+	/// get the agents properties
+	#[must_use]
+	pub fn props(&self) -> Arc<RwLock<P>> {
+		self.props.clone()
+	}
+
+	//#[cfg_attr(doc, doc(cfg(feature = "liveliness")))]
 	/// activate sending liveliness information
 	#[cfg(feature = "liveliness")]
 	pub fn liveliness(&mut self, activate: bool) {
 		self.liveliness = activate;
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "liveliness")))]
 	/// get a builder for a subscriber for the liveliness information
 	#[cfg(feature = "liveliness")]
+	#[must_use]
 	pub fn liveliness_subscriber(&self) -> LivelinessSubscriberBuilder<P> {
 		LivelinessSubscriberBuilder {
 			subscriber: self.liveliness_subscriber.clone(),
 			communicator: self.com.clone(),
 			props: self.props.clone(),
 			key_expr: None,
-			callback: None,
+			put_callback: None,
+			delete_callback: None,
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "subscriber")))]
 	/// get a builder for a Subscriber
 	#[cfg(feature = "subscriber")]
+	#[must_use]
 	pub fn subscriber(&self) -> SubscriberBuilder<P> {
 		SubscriberBuilder {
 			collection: self.subscribers.clone(),
@@ -130,8 +151,10 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "queryable")))]
 	/// get a builder for a Queryable
 	#[cfg(feature = "queryable")]
+	#[must_use]
 	pub fn queryable(&self) -> QueryableBuilder<P> {
 		QueryableBuilder {
 			collection: self.queryables.clone(),
@@ -142,8 +165,10 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "publisher")))]
 	/// get a builder for a Publisher
 	#[cfg(feature = "publisher")]
+	#[must_use]
 	pub fn publisher(&self) -> PublisherBuilder<'a, P> {
 		PublisherBuilder {
 			collection: self.publishers.clone(),
@@ -153,8 +178,10 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "query")))]
 	/// get a builder for a Query
 	#[cfg(feature = "query")]
+	#[must_use]
 	pub fn query(&self) -> QueryBuilder<P> {
 		QueryBuilder {
 			collection: self.queries.clone(),
@@ -165,12 +192,11 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "timer")))]
 	/// get a builder for a Timer
 	#[cfg(feature = "timer")]
-	pub fn timer(&self) -> TimerBuilder<P>
-	where
-		P: Default,
-	{
+	#[must_use]
+	pub fn timer(&self) -> TimerBuilder<P> {
 		TimerBuilder {
 			collection: self.timers.clone(),
 			communicator: self.com.clone(),
@@ -185,9 +211,6 @@ where
 	/// # Panics
 	///
 	pub async fn start(&mut self) {
-		// avoid warning
-		let _props = self.props.clone();
-
 		// start all registered queryables
 		#[cfg(feature = "queryable")]
 		self.queryables